Add original and custom look. #2
The game has two themes. One is orignal (the black and white game like
it used to be), the other one is custom (where I decided to have a bit
of fun). It is very easy to change between them. I plan to have a proper
menu where the user can choose between them and hit play again after the
game is over. The files are also reorganized into code and resource
directories.